refactor(ui): extract code row layout
