+++ src/ui/diff/codeRowAffordance.ts
@@ -1,0 +1,5 @@
+/** Text painted for the code-row add-note affordance. */
+export const CODE_ROW_ADD_NOTE_BADGE_TEXT = "[+]";
+
+/** Terminal columns reserved for the code-row add-note affordance. */
+export const CODE_ROW_ADD_NOTE_BADGE_WIDTH = CODE_ROW_ADD_NOTE_BADGE_TEXT.length;
